# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: openlog Query and Management API (M1)
  version: 0.2.0
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
  description: |
    Machine-readable form of [api.md](api.md). The handlers in `internal/api` are the source of truth.
    Times are accepted as RFC3339 or unix milliseconds and returned as RFC3339 with nanoseconds (UTC).
    The tenant is taken from the authenticated session or API key; no endpoint accepts a tenant parameter.
    Users in several organizations select one with the `X-Openlog-Org-Id` header.
    Cookie-authenticated requests with methods other than GET/HEAD/OPTIONS must send `X-CSRF-Token`
    (the `csrf_token` of the Me object).
servers:
  - url: /
security:
  - session: []
  - apiKey: []
...
paths:
  /api/v1/hosts:
    get:
      tags: [hosts]
      operationId: listHosts
      summary: Hosts seen in the last 24h, ordered by host_name
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [hosts]
                properties:
                  hosts:
                    type: array
                    items: { $ref: "#/components/schemas/Host" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/Internal" }
        "504": { $ref: "#/components/responses/Timeout" }
  /api/v1/hosts/{host_id}:
    get:
      tags: [hosts]
      operationId: getHost
      parameters:
        - $ref: "#/components/parameters/HostID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Host" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
        "504": { $ref: "#/components/responses/Timeout" }
...
  /api/v1/hosts/{host_id}/inventory:
    get:
      tags: [inventory]
      operationId: getHostInventory
      description: |
        Items of the host's latest complete snapshot. A known host without a complete snapshot returns 200 with
        `snapshot_id: ""`, `snapshot_time: null` and empty `items`. A host id without a host record in the
        caller's organization returns 404.
      parameters:
        - $ref: "#/components/parameters/HostID"
        - name: category
          in: query
          schema: { $ref: "#/components/schemas/InventoryCategory" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
        "504": { $ref: "#/components/responses/Timeout" }
  /api/v1/hosts/{host_id}/services:
    get:
      tags: [inventory]
      operationId: getHostServices
      description: Shortcut for `category=discovered_service`; `data` is the discovered service body. Unknown host → 404.
      parameters:
        - $ref: "#/components/parameters/HostID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
        "504": { $ref: "#/components/responses/Timeout" }
  /api/v1/inventory/search:
    get:
      tags: [inventory]
      operationId: searchInventory
      description: Items in `category` whose key contains `q` (case-insensitive) across all hosts' latest snapshots, ordered by key then host.
      parameters:
        - name: category
          in: query
          required: true
          schema: { $ref: "#/components/schemas/InventoryCategory" }
        - name: q
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/InventorySearchItem" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/Internal" }
        "504": { $ref: "#/components/responses/Timeout" }
  /api/v1/logs:
    get:
      tags: [logs]
      operationId: listLogs
      description: |
        Excludes inventory events. `q` is a case-insensitive substring match on body. Newest first.
        `attr.<key>=<value>` parameters are exact-match filters on log record attributes; only the keys listed
        below are accepted (any other `attr.*` key, an empty or repeated value, or a value over 1024 bytes → 400).
        Infra agent log records have an empty `service_name`; filter them with `host_id` and these attributes.
      parameters:
        - name: host_id
          in: query
          schema: { type: string }
        - name: service
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - name: severity_min
          in: query
          description: Severity number 1-24 or a name (TRACE, DEBUG, INFO, WARN, WARNING, ERROR, FATAL).
          schema: { type: string, example: WARN }
        - name: trace_id
          in: query
          schema: { type: string }
        - name: attr.openlog.log.source
          in: query
          description: "`file` or `journald`"
          schema: { type: string, example: file }
        - name: attr.log.file.path
          in: query
          schema: { type: string, example: /var/log/nginx/access.log }
        - name: attr.log.file.name
          in: query
          schema: { type: string, example: access.log }
        - name: attr.openlog.discovery.id
          in: query
          description: Discovery rule id of the service whose log glob matched the file.
          schema: { type: string, example: nginx }
        - name: attr.openlog.systemd.unit
          in: query
          schema: { type: string, example: nginx.service }
        - name: attr.openlog.syslog.identifier
          in: query
          schema: { type: string, example: sshd }
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [logs]
                properties:
                  logs:
                    type: array
                    items: { $ref: "#/components/schemas/LogRecord" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/Internal" }
        "504": { $ref: "#/components/responses/Timeout" }
  /api/v1/traces/{trace_id}:
    get:
      tags: [traces]
      operationId: getTrace
      parameters:
        - name: trace_id
          in: path
          required: true
          description: 32 hex characters (case-insensitive).
          schema: { type: string, pattern: "^[0-9a-fA-F]{32}$" }
      responses:
        "200":
          description: All spans of the trace ordered by start time
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Trace" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
        "504": { $ref: "#/components/responses/Timeout" }

  # ---------------------------------------------------------------- apm (apm.md)
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: openlog_session
      description: Session cookie from POST /api/v1/auth/login. Mutating requests also need the X-CSRF-Token header.
    apiKey:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer ola_…` — read-only API key (acts as viewer)."
    licenseKey:
      type: apiKey
      in: header
      name: openlog-license-key
      description: OPENLOG_AUTH_MODE=static only (development/tests); also accepted as a bearer token in that mode.
  parameters:
    HostID:
      name: host_id
      in: path
      required: true
      schema: { type: string }
    From:
      name: from
      in: query
      description: RFC3339 or unix milliseconds. Default now − 1h.
      schema: { type: string }
    To:
      name: to
      in: query
      description: RFC3339 or unix milliseconds. Default now.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      description: Default 100, capped by OPENLOG_API_MAX_ROWS.
      schema: { type: integer, minimum: 1 }
...
  responses:
    BadRequest:
      description: invalid_argument
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthenticated:
      description: unauthenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: not_found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Internal:
      description: internal
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Timeout:
      description: timeout
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_argument, unauthenticated, permission_denied, not_found, already_exists, failed_precondition, resource_exhausted, internal, unavailable, timeout]
            message: { type: string }
    Timestamp:
      type: string
      format: date-time
      description: RFC3339 with nanoseconds, UTC (e.g. 2026-09-13T10:00:00.000000000Z)
    StringMap:
      type: object
      additionalProperties: { type: string }
    Host:
      type: object
      required: [host_id, host_name, os_description, arch, agent_version, last_seen, resource_attributes]
      properties:
        host_id: { type: string }
        host_name: { type: string }
        os_description: { type: string }
        arch: { type: string }
        agent_version: { type: string }
        last_seen: { $ref: "#/components/schemas/Timestamp" }
        resource_attributes: { $ref: "#/components/schemas/StringMap" }
...
    InventoryCategory:
      type: string
      description: Known categories (semantic-conventions §3.3); the API accepts any string.
      examples: [os, hardware, kernel_module, package, systemd_unit, listening_port, process, user, network_interface, mount, discovered_service]
    InventoryItem:
      type: object
      required: [category, key, data]
      properties:
        category: { type: string }
        key: { type: string }
        data:
          description: The category's JSON body; non-JSON bodies are returned as a string.
    InventoryResponse:
      type: object
      required: [snapshot_id, snapshot_time, items]
      properties:
        snapshot_id: { type: string }
        snapshot_time:
          oneOf:
            - { $ref: "#/components/schemas/Timestamp" }
            - { type: "null" }
        items:
          type: array
          items: { $ref: "#/components/schemas/InventoryItem" }
    InventorySearchItem:
      allOf:
        - $ref: "#/components/schemas/InventoryItem"
        - type: object
          required: [host_id]
          properties:
            host_id: { type: string }
            host_name:
              type: string
              description: Omitted when the host is not found in the hosts table.
...
    LogRecord:
      type: object
      required: [timestamp, severity_text, severity_number, body, host_id, service_name, trace_id, span_id, attributes, resource_attributes]
      properties:
        timestamp: { $ref: "#/components/schemas/Timestamp" }
        severity_text: { type: string }
        severity_number: { type: integer, minimum: 0, maximum: 255 }
        body: { type: string }
        host_id: { type: string }
        service_name: { type: string }
        trace_id: { type: string, description: "32 lowercase hex characters or empty" }
        span_id: { type: string, description: "16 lowercase hex characters or empty" }
        attributes: { $ref: "#/components/schemas/StringMap" }
        resource_attributes: { $ref: "#/components/schemas/StringMap" }
    SpanKind:
      type: string
      enum: [unspecified, internal, server, client, producer, consumer]
    SpanStatusCode:
      type: string
      enum: [unset, ok, error]
    SpanEvent:
      type: object
      required: [timestamp, name, attributes]
      properties:
        timestamp: { type: string, description: "RFC3339 with nanoseconds; empty if missing" }
        name: { type: string }
        attributes: { $ref: "#/components/schemas/StringMap" }
    Span:
      type: object
      required: [span_id, parent_span_id, name, kind, service_name, start, duration_ns, status_code, status_message, attributes, resource_attributes, events]
      properties:
        span_id: { type: string }
        parent_span_id: { type: string, description: "Empty for root spans" }
        name: { type: string }
        kind: { $ref: "#/components/schemas/SpanKind" }
        service_name: { type: string }
        start: { $ref: "#/components/schemas/Timestamp" }
        duration_ns: { type: integer, minimum: 0 }
        status_code: { $ref: "#/components/schemas/SpanStatusCode" }
        status_message: { type: string }
        attributes: { $ref: "#/components/schemas/StringMap" }
        resource_attributes: { $ref: "#/components/schemas/StringMap" }
        events:
          type: array
          items: { $ref: "#/components/schemas/SpanEvent" }
    Trace:
      type: object
      required: [trace_id, spans]
      properties:
        trace_id: { type: string }
        spans:
          type: array
          items: { $ref: "#/components/schemas/Span" }

    # ------------------------------------------------------------ auth & management
    # ---- APM (apm.md) ----
...